+++ src/agent_behaviours/sensors/angle_vector3.h
@@ -12,11 +12,11 @@ class UtilityAIAngleVector3Sensor : public UtilityAISensor {
 private:
     Vector3 _from_vector;
     Vector3 _to_vector;
-    double  _angle_radian;
-    double  _angle_degrees;
-    double  _max_angle_radian; 
-    double  _max_angle_degrees; 
-    double  _one_over_max_angle_radian;
+    float  _angle_radian;
+    float  _angle_degrees;
+    float  _max_angle_radian; 
+    float  _max_angle_degrees; 
+    float  _one_over_max_angle_radian;
 
 protected:
     static void _bind_methods();
@@ -27,7 +27,7 @@ public:
     
     // Handling functions. 
 
-    virtual double evaluate_sensor_value() override;
+    virtual float evaluate_sensor_value() override;
 
     // Getters and setters for attributes.
     
@@ -39,19 +39,19 @@ public:
     void set_from_vector3( Vector3 from );
     Vector3 get_from_vector3() const;
 
-    void set_max_angle_radian( double max_angle_radian );
-    double get_max_angle_radian() const;
+    void set_max_angle_radian( float max_angle_radian );
+    float get_max_angle_radian() const;
 
-    void set_max_angle_degrees( double max_angle_radian );
-    double get_max_angle_degrees() const;
+    void set_max_angle_degrees( float max_angle_radian );
+    float get_max_angle_degrees() const;
 
     // Debugging / current values.
     
-    void set_angle_radian( double angle_radian );
-    double get_angle_radian() const;
+    void set_angle_radian( float angle_radian );
+    float get_angle_radian() const;
 
-    void set_angle_degrees( double angle_degrees );
-    double get_angle_degrees() const;
+    void set_angle_degrees( float angle_degrees );
+    float get_angle_degrees() const;
 
 };
